requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-deepseek==0.1.1
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
python-dotenv==1.0.1
pydantic==2.10.6
pyyaml==6.0.2
gunicorn==21.2.0
pytest==7.4.3